Split errors.ts into errors/ folder (one class per file)
Replaces the monolithic `src/errors.ts` (which required a file-level
`eslint-disable max-classes-per-file`) with a dedicated `src/errors/`
folder containing one file per error class:

- `errors/base.ts`            — `MelCloudError` + `isMelCloudError`
- `errors/authentication.ts`  — `AuthenticationError`
- `errors/rate-limit.ts`      — `RateLimitError`
- `errors/transient-server.ts` — `TransientServerError`
- `errors/network.ts`         — `NetworkError`
- `errors/index.ts`           — barrel re-export

This eliminates the last `eslint-disable max-classes-per-file`
comment. Each file now contains exactly one class — no lint
overrides needed.

All consumers import from `errors/index.ts` via the existing barrel
pattern. Public API surface (main.ts re-exports) unchanged.

501 tests passing, 100% coverage.
