+++ src/errors/index.ts
@@ -1,0 +1,5 @@
+export { AuthenticationError } from './authentication.ts'
+export { isMelCloudError, MelCloudError } from './base.ts'
+export { NetworkError } from './network.ts'
+export { RateLimitError } from './rate-limit.ts'
+export { TransientServerError } from './transient-server.ts'
