+++ src/resilience/index.ts
@@ -21,5 +21,5 @@ export {
   NetworkError,
   RateLimitError,
   TransientServerError,
-} from '../errors.ts'
+} from '../errors/index.ts'
 export { areaId, buildingId, deviceId, floorId } from '../types/ids.ts'
